将获取用户信息的路径放到 /users 下
