+++ routes/auth.ts
@@ -2,10 +2,7 @@ import express from 'express'
 import {encryptPassword, generateJWT} from '@/utils/auth'
 import passport from '@/plugins/passport'
 import UserModel from '@/models/UserModel'
-import JobModel from '@/models/JobModel'
 import {v4 as uuidv4} from 'uuid'
-import ReferModel from '@/models/ReferModel'
-import ResumeModel from '@/models/ResumeModel'
 
 // '/auth'
 const AuthRouter = express.Router()
@@ -54,57 +51,6 @@ AuthRouter.post('/register', async (req, res) => {
   })
 
   res.json(user)
-})
-
-// 获取个人信息
-AuthRouter.get('/user', passport.authenticate('jwt', {session: false}), async (req, res) => {
-  const {userId} = req.user as TJWTUser
-
-  const user = await UserModel.findByPk(userId, {
-    include: [JobModel, {
-      model: ResumeModel,
-      as: 'resumeList',
-      limit: 1,
-      order: [['createdAt', 'DESC']]
-    }],
-  })
-
-  if (!user) {
-    res.status(404)
-    return res.json({message: '该用户不存在'})
-  }
-
-  // 统计已处理自己的 refer
-  const myReferTotal = await ReferModel.count({
-    where: {refereeId: userId}
-  })
-  const approvedMyReferCount = await ReferModel.count({
-    where: {refereeId: userId, status: 'approved'}
-  })
-  const otherReferTotal = await ReferModel.count({
-    where: {refererId: userId}
-  })
-  const approvedOtherReferCount = await ReferModel.count({
-    where: {refererId: userId, status: 'approved'}
-  })
-
-  const {job, resumeList, ...userInfo}: any = user.toJSON()
-  const info: TUser = {
-    ...userInfo,
-    myReferTotal,
-    approvedMyReferCount,
-    otherReferTotal,
-    approvedOtherReferCount
-  }
-
-  res.json({
-    info,
-    job: {
-      ...job,
-      requiredFields: job.requiredFields.split(',')
-    },
-    resume: resumeList.length > 0 ? resumeList[0] : null,
-  })
 })
 
 export default AuthRouter
